[refactor] -updated folder structure for nav
